Revert incorrect change in per_epoch_processing and check progressive balances cache in fork choice.
