+++ consensus/state_processing/src/per_epoch_processing/altair/justification_and_finalization.rs
@@ -3,11 +3,9 @@ use crate::per_epoch_processing::Error;
 use crate::per_epoch_processing::{
     weigh_justification_and_finalization, JustificationAndFinalizationState,
 };
-use crate::EpochProcessingError;
 use safe_arith::SafeArith;
-use types::beacon_state::Error as BeaconStateError;
 use types::consts::altair::TIMELY_TARGET_FLAG_INDEX;
-use types::{BeaconState, EthSpec, ProgressiveTotalBalances};
+use types::{BeaconState, EthSpec};
 
 /// Update the justified and finalized checkpoints for matching target attestations.
 pub fn process_justification_and_finalization<T: EthSpec>(
@@ -20,43 +18,6 @@ pub fn process_justification_and_finalization<T: EthSpec>(
         return Ok(justification_and_finalization_state);
     }
 
-    // Use the balances from the `ProgressiveTotalBalances` cache if available, otherwise calculate them.
-    let (total_active_balance, previous_target_balance, current_target_balance) =
-        get_progressive_total_balances(state)
-            .or_else(|_| calculate_total_balances(state, participation_cache))?;
-
-    weigh_justification_and_finalization(
-        justification_and_finalization_state,
-        total_active_balance,
-        previous_target_balance,
-        current_target_balance,
-    )
-}
-
-fn get_progressive_total_balances<T: EthSpec>(
-    state: &BeaconState<T>,
-) -> Result<(u64, u64, u64), EpochProcessingError> {
-    let (_, total_active_balance) = state
-        .total_active_balance()
-        .ok_or(BeaconStateError::TotalActiveBalanceCacheUninitialized)?;
-
-    let progressive_total_balances: &ProgressiveTotalBalances = state.progressive_total_balances();
-    let (previous_target_balance, current_target_balance) = (
-        progressive_total_balances.previous_epoch_target_attesting_balance()?,
-        progressive_total_balances.current_epoch_target_attesting_balance()?,
-    );
-
-    Ok((
-        total_active_balance,
-        previous_target_balance,
-        current_target_balance,
-    ))
-}
-
-fn calculate_total_balances<T: EthSpec>(
-    state: &BeaconState<T>,
-    participation_cache: &ParticipationCache,
-) -> Result<(u64, u64, u64), EpochProcessingError> {
     let previous_epoch = state.previous_epoch();
     let current_epoch = state.current_epoch();
     let previous_indices = participation_cache
@@ -66,9 +27,10 @@ fn calculate_total_balances<T: EthSpec>(
     let total_active_balance = participation_cache.current_epoch_total_active_balance();
     let previous_target_balance = previous_indices.total_balance()?;
     let current_target_balance = current_indices.total_balance()?;
-    Ok((
+    weigh_justification_and_finalization(
+        justification_and_finalization_state,
         total_active_balance,
         previous_target_balance,
         current_target_balance,
-    ))
+    )
 }
